idlestat: run a command with idlestat
Add the support to run a command line with idlestat:

./idlestat -o out.dat -- /bin/sleep 3

Without the '-t' option, idlestat will blindly use the current buffer
size which may not fit with the 'command' duration and lead to a buffer
overflow.

With the example, it would make sense to set the buffer size:

./idlestat -o out.dat -t 3 -- /bin/sleep 3

This is a workaround, until we find a way to set the buffer size automatically.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>