Fix bad check with waitpid.
With the current code, when the child exits on timeout, waitpid returns -1.

The check right after will kill the process without waiting it, so none of the
macros after will conclude the process exited correctly leading to the end of
the function with a return -1.

Kill the process in case of timeout and wait it.

We don't care of the other errors because waitpid returns ECHLD, EINVAL.

 * ECHLD : we won't wait twice our child
 * EINVAL: we are not passing flag

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>