Allow idlestat to parse older versions of trace output
Re-work TRACE_FORMAT to allow idlestat to parse older kernels
trace output. This has been tested successfully on Debian 7.5
with a 3.2.x kernel and also on 3.15 kernel.

And example of the the newer format is:
 <idle>-0     [002] .N..  1420.031852: cpu_idle: state=4294967295 cpu_id=2

where as the older format is:

<idle>-0     [000]    91.365043: cpu_idle: state=1 cpu_id=0

The modified TRACE_FORMAT will handle both inputs by consuming characters
upto the floating point time field.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>