pstate_init: re-factor idlestat_wake_all to support P-State injection
We need to inject an FTrace marker for each CPU which reports the current
P-State assigned by CPUFreq.

This patch removes a little run-time optimisation which would make
unnecessary complex the code just to handle the marker injection in the
current CPU. Since that method is called just on the setup step, that
little optimisation could be avoided in favour of a more readable code.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>