pstate_init: ignore intial CPUFreq/CPUIdle events
All the events reported by CPUFreq and CPUIdle before the PState marker
has been found are not ignored by idletrace.

The idea is that idlestat starts to trace events only after a PState marker,
which is however forced on each CPU by idlestat_wake_all.
When such a marker is injectted we know for sure:
1. that the CPU is running, in the PState reporetd by the marker
2. the CPU will enter an CState right after the marker has been injected

The point 2 is not ensured, but quite likely at least on lightly loaded systems.
In any case, any further C-State event reported by CPUIdle will be for sure
the first event tracked by idlestat.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>