scenario: dump topology info as comments
This enables the usage of sythetics scenarios in idlestat.

All the info put in the header section of the output trace are now dumped
with a '#' prefix.
A further modification, which ignores comments and empty lines, it allows
to use a trace generated by idlestat as an input trace for the same tool.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>