psci_proxy 1/4: update data structures
This introduces a new set of per-CPU fields which allows to better track
the single CPU status while avoiding overlapping of meaning with different
values used for the same field.

Now, it is more-or-less a set of flags which could be optimises in a following
version by using a bitfield to track them all thus reducing memory footprint.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>