hwmon: add support to dump estimated/actual energy
A set of methods and data structures have been added which allow to
append on the tc2_energy.dat output file the amount of energy estimated
by the model, as well as the actual energy measured via HWMON.
These values are reported per cluster with a breakdown per Cap, Idle and
Wakeups energy contribution for the model predicted energy consumptions.

The output file is suitable to be parsed and plotted using a script
provided by one of the following patches.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>