trace_trimming: update intersection code to consider marker timestamps
The START and END timestamp could be used to discard FTrace events
not generated while the workload was in execution, e.g. initial spurious
events.

This patch setup the required support to pass the start end end timestamp
of the workload execution to the intersection computation code.
From here they could be exploited to trim/discard events fall (partially)
outside of the experiment execution timeframe.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>