Fix CPU affinity mask
To properly identify the initial PState of each CPU, the idlestat_wake_all
function is used to force an idle state exit on all CPUs by pinning idlestat
on each CPUs. However, this routine forgets to recover the original mask
at the end of the pinning process.

This patch solve this issue, by properly recovering the original CPU
affinity mask. Moreover, it avoids to wake up CPUs that will not be used
to run the workload.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Acked-by: Larry Bassel <larry.bassel@linaro.org>
Tested-by: Larry Bassel <larry.bassel@linaro.org>