Add 'patches' directory to the ignored list of files in .gitignore
The quilt tool is still widely used in conjunction with git, it needs a
specific directory, by convention 'patches' to store the patches for the
serie.

Add this directory in the .gitignore file, so we are no longer annoyed by this
directory when we check the status of git reporting it as an untracked file.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Acked-by: Larry Bassel <larry.bassel@linaro.org>
Tested-by: Larry Bassel <larry.bassel@linaro.org>