idlestat: Add -o option to save output report to a file
Currently the serial terminal connected to the boards running idlestat are
restricted to be at least 80 characters wide to output the report. Otherwise
idlestat quits with message "The terminal must be at least 80 columns wide".

Changes v1 to v2:
* Duplicate opened report file to stdout instead of replacing all printf()

Fix it by adding a "-o" option to save report output to a file.

Signed-off-by: Pi-Cheng Chen <pi-cheng.chen@linaro.org>
[amit.kucheria@linaro.org: Removed extra examples and check for bad filenames]
Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>